separate out context providers in App from the consumers in Main
